Explicit cast iterator difference
explicitly cast the result of iterator calculation (difference between
two iterators) to an unsigned int before assigning to the slide index,
which is an unsigned int.

+++ Splashentation/Standard.cpp
@@ -298,7 +298,7 @@ void Splashentation::t_play()
 			else
 			{
 				std::lock_guard<std::mutex> lockGuard(m_informationMutex);
-				m_currentSlideIndex = currentSlide - m_slides.begin();
+				m_currentSlideIndex = static_cast<unsigned int>(currentSlide - m_slides.begin());
 			}
 		}
 	}
